+++ evaluator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.3](https://github.com/kguardian-dev/kguardian/compare/evaluator/v0.3.2...evaluator/v0.3.3) (2026-07-23)
+
+
+### Bug Fixes
+
+* **deps:** update kubernetes monorepo to v0.36.3 ([#1145](https://github.com/kguardian-dev/kguardian/issues/1145)) ([0d8b8cc](https://github.com/kguardian-dev/kguardian/commit/0d8b8cc06326656a11edeb7f5bfdaa67d84d09f5))
+
 ## [0.3.2](https://github.com/kguardian-dev/kguardian/compare/evaluator/v0.3.1...evaluator/v0.3.2) (2026-06-29)